streamlit_app/requirements.txt
streamlit==1.36.0
pandas>=2.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
plotly>=5.20.0
python-dotenv>=1.0.1